Pendency APis

+++ rule_engine/rcm/src/main/java/com/tricon/rcm/dto/customquery/FreshClaimDataDto.java
@@ -18,8 +18,8 @@ public interface FreshClaimDataDto {
 	String getLastTeam();
 	int getClaimAge();
 	String getTimelyFilingLimitData();
-	double getBilledAmount();
-	double getPrimTotal();
-	double getSecTotal();
+	float getBilledAmount();
+	float getPrimTotal();
+	float getSecTotal();
 	
 }
